refactor: extract HeaderCell to shared table components (DRY)
