+++ src/renderer/components/index.ts
@@ -1,3 +1,4 @@
 export { MainLayout } from './MainLayout';
 export { BackgroundLayer } from './BackgroundLayer';
 export { SectionHeading } from './SectionHeading';
+export { HeaderCell } from './table';
