+++ src/renderer/components/table/index.ts
@@ -1,0 +1,1 @@
+export { HeaderCell } from './HeaderCell';
