+++ src/renderer/content/Championship.tsx
@@ -1,5 +1,5 @@
 import { useDerivedGameState } from '../hooks';
-import { SectionHeading } from '../components';
+import { SectionHeading, HeaderCell } from '../components';
 import {
   TABLE_CELL_BASE,
   TABLE_HEADER_CLASS,
@@ -25,28 +25,6 @@ const POSITION_COL_CLASS = 'w-16';
 // ===========================================
 // TABLE COMPONENTS
 // ===========================================
-
-type TextAlign = 'left' | 'center' | 'right';
-
-const TEXT_ALIGN_CLASSES: Record<TextAlign, string> = {
-  left: 'text-left',
-  center: 'text-center',
-  right: 'text-right',
-};
-
-interface HeaderCellProps {
-  children: React.ReactNode;
-  align?: TextAlign;
-  className?: string;
-}
-
-function HeaderCell({ children, align = 'center', className = '' }: HeaderCellProps) {
-  return (
-    <th className={`${TABLE_CELL_BASE} ${TEXT_ALIGN_CLASSES[align]} ${className}`.trim()}>
-      {children}
-    </th>
-  );
-}
 
 interface StatCellProps {
   value: number;
